Add files via upload
The vocabulary is created from searching words throught the input field under SEARCH ONLINE DICTIONARIES. The program works as follows: 1) Copy any german texts and paste it in the 'Study Text Box', and save it in the folder and name of your choice. 2) The same German text paste in DeepL.com/translator for a quality translation.  3) Start studying the german text and search for unknown words as mentioned above. 4) Save your vocabulary to study and learn the words and expressions later. You can check your knowledge of the vocabulary by taking a test as follows: a) Choose a File and type your answer in the input field below the Multiline text box on the right (in the middle). Press ENTER to check your answer. Press 'New Word' to continue. NOTE: When learning a foreign language, we usually need example sentences to better understand the use of a verb (mainly). Type a verb in the input field under SEARCH ONLINE DICTIONARIES and then click the button 'Collins-Examples' to search for sentences that examplify the use of the verb. NOTE ALSO: The 'Flip Sentences' button serves to test your knowledge of words / expressions from the English to German side.  In other words, the German to English sentences can be flipped by 180 degrees.